moved protocol callbacks into the proper place where they are used: embObjLib
