+++ src/libraries/icubmod/embObjLib/embObjLib.cmake
@@ -8,7 +8,7 @@ else()
     set(EO_BASE_DIR         ${CMAKE_SOURCE_DIR}/src/libraries/icubmod/embObjLib/embObj/)
 endif()
 
-set(CORE_FOLDER  		${EO_BASE_DIR}/embobj/core/core)
+set(CORE_FOLDER         ${EO_BASE_DIR}/embobj/core/core)
 set(CORE_YEE            ${EO_BASE_DIR}/embobj/core/exec/yarp/)
 
 
@@ -39,7 +39,7 @@ set(PATH_TO_CALLBACK ${CMAKE_SOURCE_DIR}/src/libraries/icubmod/)
     set(NVS_MC_CBK_FOLDER       ${PATH_TO_CALLBACK}/embObjMotionControl/usrcbk/)
     
     set(NVS_MNGMNT_SRC_FOLDER   ${PROT_FOLDER}/protocol/src/)
-    set(NVS_MNGMNT_CBK_FOLDER   ${PATH_TO_CALLBACK}/embObjLib/usrcbk/)
+    set(NVS_CBK_FOLDER          ${PATH_TO_CALLBACK}/embObjLib/usrcbk/)
     
     set(NVS_SENSORS_SRC_FOLDER  ${PROT_FOLDER}/protocol/src/)
     set(NVS_SENSORS_CBK_FOLDER  ${PATH_TO_CALLBACK}/embObjAnalog/usrcbk/)
@@ -70,7 +70,7 @@ set(PATH_TO_CALLBACK ${CMAKE_SOURCE_DIR}/src/libraries/icubmod/)
                                 ${ICUB_INTERFACE_CORE_FOLDER}
                                 ${ICUB_INTERFACE_CORE_FOLDER}/robotInterface
                                 ${NVS_SKIN_SRC_FOLDER}
-                                ${NVS_MNGMNT_CBK_FOLDER}
+                                ${NVS_CBK_FOLDER}
                                 ${NVS_MC_CBK_FOLDER}
                                 ${NVS_SKIN_CBK_FOLDER}
                                 ${NVS_SENSORS_CBK_FOLDER}
