+++ src/libraries/icubmod/embObjLib/usrcbk/EoProtocolAS_fun_userdef.c
@@ -17,6 +17,11 @@
 */
 
 
+/*  @file       EOProtocolAS_fun_userdef.c
+    @brief      This file keeps callbacks used for AS protocol in icub-main
+    @author     marco.accame@iit.it
+    @date       22 mar 2016
+**/
 
 // --------------------------------------------------------------------------------------------------------------------
 // - external dependencies
@@ -24,17 +29,17 @@
 
 
 #include "EoCommon.h"
-#include "EOnv.h"
 #include "EoProtocol.h"
+#include "EOnv.h"
+#include "EOarray.h"
 
-#include <FeatureInterface.h>
-
+#include "FeatureInterface.h"
 
 // --------------------------------------------------------------------------------------------------------------------
 // - declaration of extern public interface
 // --------------------------------------------------------------------------------------------------------------------
 
-#include "EoProtocolMC.h"
+#include "EoProtocolAS.h"
 
 
 
@@ -43,12 +48,14 @@
 // --------------------------------------------------------------------------------------------------------------------
 // empty-section
 
+
 // --------------------------------------------------------------------------------------------------------------------
 // - declaration of static functions
 // --------------------------------------------------------------------------------------------------------------------
 
-static void wake(const EOnv* nv);
+static void handle_data_analogarray(const EOnv* nv, const eOropdescriptor_t* rd);
 
+static void handle_data_inertial(const EOnv* nv, const eOropdescriptor_t* rd);
 
 // --------------------------------------------------------------------------------------------------------------------
 // - definition (and initialisation) of static variables
@@ -61,23 +68,20 @@ static void wake(const EOnv* nv);
 // - definition of extern public functions
 // --------------------------------------------------------------------------------------------------------------------
 
-extern void eoprot_fun_ONSAY_mc(const EOnv* nv, const eOropdescriptor_t* rd)
+
+extern void eoprot_fun_ONSAY_as(const EOnv* nv, const eOropdescriptor_t* rd)
 {
     // marco.accame on 18 mar 2014: this function is called when a say<id32, data> rop is received
-    // and the id32 is about the motion control endpoint. this function is common to every board.
+    // and the id32 is about the analog sensors endpoint. this function is common to every board.
     // it is used this function and not another one because inside the hostTransceiver object it was called:
-    // eoprot_config_onsay_endpoint_set(eoprot_endpoint_motioncontrol, eoprot_fun_ONSAY_mc);
+    // eoprot_config_onsay_endpoint_set(eoprot_endpoint_analogsensors, eoprot_fun_ONSAY_as);
 
     // the aim of this function is to wake up a thread which is blocked because it has sent an ask<id32>
-    // the wake up funtionality is implemented in two modes, depending on the wait mechanism used:
-    // a. in initialisation, embObjMotionControl sets some values and then reads them back.
+    // the wake up funtionality is implemented in one mode only:
+    // a. in initialisation, embObjAnalogSensor sets some values and then reads them back.
     //    the read back sends an ask<id32, signature=0xaa000000>. in such a case the board sends back
     //    a say<id32, data, signature = 0xaa000000>. thus, if the received signature is 0xaa000000, then
     //    we must unblock using feat_signal_network_reply().
-    // b. during runtime, some methods send a blocking ask<id32> without signature. It is the case of instance
-    //    of getPidRaw() which waits with a eoThreadEntry::synch() call. in such a case the board send back a
-    //    normal say<id32, data> with nos signature. in this case we unlock with wake().
-
 
     if(0xaa000000 == rd->signature)
     {   // case a:
@@ -86,67 +90,69 @@ extern void eoprot_fun_ONSAY_mc(const EOnv* nv, const eOropdescriptor_t* rd)
             char str[256] = {0};
             char nvinfo[128];
             eoprot_ID2information(rd->id32, nvinfo, sizeof(nvinfo));
-            snprintf(str, sizeof(str), "eoprot_fun_ONSAY_mc() received an unexpected message w/ 0xaa000000 signature for %s", nvinfo);
+            snprintf(str, sizeof(str), "eoprot_fun_ONSAY_as() received an unexpected message w/ 0xaa000000 signature for %s", nvinfo);
             feat_PrintWarning(str);
             return;
         }
     }
-    else
-    {   //case b:
-        wake(nv);
-    }
 }
 
 
-extern void eoprot_fun_UPDT_mc_joint_status_core(const EOnv* nv, const eOropdescriptor_t* rd)
+extern void eoprot_fun_UPDT_as_strain_status_calibratedvalues(const EOnv* nv, const eOropdescriptor_t* rd)
 {
-    feat_manage_motioncontrol_data(eo_nv_GetIP(nv), rd->id32, (void *)rd->data);
+    handle_data_analogarray(nv, rd);
 }
 
 
-extern void eoprot_fun_UPDT_mc_motor_status_basic(const EOnv* nv, const eOropdescriptor_t* rd)
+extern void eoprot_fun_UPDT_as_strain_status_uncalibratedvalues(const EOnv* nv, const eOropdescriptor_t* rd)
 {
+    handle_data_analogarray(nv, rd);
 }
 
 
-extern void eoprot_fun_UPDT_mc_joint_status(const EOnv* nv, const eOropdescriptor_t* rd)
+extern void eoprot_fun_UPDT_as_mais_status_the15values(const EOnv* nv, const eOropdescriptor_t* rd)
 {
-    feat_manage_motioncontrol_data(eo_nv_GetIP(nv), rd->id32, (void *)rd->data);
+    handle_data_analogarray(nv, rd);
 }
+
+
+extern void eoprot_fun_UPDT_as_inertial_status(const EOnv* nv, const eOropdescriptor_t* rd)
+{
+    handle_data_inertial(nv, rd);
+}
+
+//extern void eoprot_fun_UPDT_as_inertial_status_accelerometer(const EOnv* nv, const eOropdescriptor_t* rd)
+//{
+//    handle_data_inertial_acc(nv, rd);
+//}
+
+//extern void eoprot_fun_UPDT_as_inertial_status_gyroscope(const EOnv* nv, const eOropdescriptor_t* rd)
+//{
+//    handle_data_inertial_gyr(nv, rd);
+//}
+
+
 
 
 // --------------------------------------------------------------------------------------------------------------------
-// - definition of static functions
+// - definition of static functions 
 // --------------------------------------------------------------------------------------------------------------------
 
-
-static void wake(const EOnv* nv)
+static void handle_data_analogarray(const EOnv* nv, const eOropdescriptor_t* rd)
 {
-    eOprotID32_t id32 = 0;
-    eOprotProgNumber_t prognum = 0 ;
-    void *mchandler = (void*) feat_MC_handler_get(eo_nv_GetIP(nv), eo_nv_GetID32(nv));
-    if(NULL == mchandler)
+    EOarray* arrayof = (EOarray*)rd->data;
+    uint8_t sizeofarray = eo_array_Size(arrayof);
+    if(0 != sizeofarray)
     {
-        printf("eoMC class not found\n");
-        return;
+        feat_manage_analogsensors_data(eo_nv_GetIP(nv), rd->id32, (void *)arrayof);
     }
-
-    id32 = eo_nv_GetID32(nv);
-    prognum = eoprot_endpoint_id2prognum(eo_nv_GetBRD(nv), id32);
-    if(eobool_false == feat_MC_mutex_post(mchandler, prognum) )
-    {
-        char nvinfo[128];
-        char ipinfo[20];
-        char str[256] = {0};
-        eoprot_ID2information(id32, nvinfo, sizeof(nvinfo));
-        eo_common_ipv4addr_to_string(eo_nv_GetIP(nv), ipinfo, sizeof(ipinfo));
-        snprintf(str, sizeof(str),"while releasing mutex for IP %s and NV %s", ipinfo, nvinfo);
-        feat_PrintWarning(str);
-    }
-
 }
 
-
+static void handle_data_inertial(const EOnv* nv, const eOropdescriptor_t* rd)
+{
+    eOas_inertial_status_t *inertialstatus  = (eOas_inertial_status_t*)rd->data;
+    feat_manage_analogsensors_data(eo_nv_GetIP(nv), rd->id32, (void *)inertialstatus);
+}
