+++ src/libraries/icubmod/embObjLib/usrcbk/EoProtocolMC_fun_userdef.c
@@ -1,7 +1,7 @@
 /*
  * Copyright (C) 2011 Department of Robotics Brain and Cognitive Sciences - Istituto Italiano di Tecnologia
- * Author:  Valentina Gaggero
- * email:   valentina.gaggero@iit.it
+ * Author:  Marco Accame
+ * email:   marco.accame@iit.it
  * website: www.robotcub.org
  * Permission is granted to copy, distribute, and/or modify this program
  * under the terms of the GNU General Public License, version 2 or any
@@ -14,33 +14,34 @@
  * WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
  * Public License for more details
- */
+*/
 
-/* @file       eOcfg_nvsEP_sk_usrcbk_ebx.c
-    @brief      This file keeps the user-defined functions used in every ems board ebx for skin endpoint
-    @author     valentina.gaggero@iit.it
-    @date       05/04/2012
- **/
 
+/*  @file       EOProtocolMC_fun_userdef.c
+    @brief      This file keeps callbacks used for MC protocol in icub-main
+    @author     marco.accame@iit.it
+    @date       22 mar 2016
+**/
 
 
 // --------------------------------------------------------------------------------------------------------------------
 // - external dependencies
 // --------------------------------------------------------------------------------------------------------------------
 
-#include "EoCommon.h"
-#include "EoCommon.h"
-#include "EoProtocol.h"
-#include "EOnv.h"
 
-#include "FeatureInterface.h"
+#include "EoCommon.h"
+#include "EOnv.h"
+#include "EoProtocol.h"
+
+#include <FeatureInterface.h>
 
 
 // --------------------------------------------------------------------------------------------------------------------
 // - declaration of extern public interface
 // --------------------------------------------------------------------------------------------------------------------
 
-#include "EoProtocolSK.h"
+#include "EoProtocolMC.h"
+
 
 
 // --------------------------------------------------------------------------------------------------------------------
@@ -49,34 +50,40 @@
 // empty-section
 
 // --------------------------------------------------------------------------------------------------------------------
-// - definition (and initialisation) of static variables
-// --------------------------------------------------------------------------------------------------------------------
-// empty-section
-
-
-// --------------------------------------------------------------------------------------------------------------------
 // - declaration of static functions
 // --------------------------------------------------------------------------------------------------------------------
-// empty-section
+
+static void wake(const EOnv* nv);
 
 
 // --------------------------------------------------------------------------------------------------------------------
-// - declaration of extern functions
+// - definition (and initialisation) of static variables
+// --------------------------------------------------------------------------------------------------------------------
+// empty section
+
+
+
+// --------------------------------------------------------------------------------------------------------------------
+// - definition of extern public functions
 // --------------------------------------------------------------------------------------------------------------------
 
-extern void eoprot_fun_ONSAY_sk(const EOnv* nv, const eOropdescriptor_t* rd)
+extern void eoprot_fun_ONSAY_mc(const EOnv* nv, const eOropdescriptor_t* rd)
 {
     // marco.accame on 18 mar 2014: this function is called when a say<id32, data> rop is received
-    // and the id32 is about the analog sensors endpoint. this function is common to every board.
+    // and the id32 is about the motion control endpoint. this function is common to every board.
     // it is used this function and not another one because inside the hostTransceiver object it was called:
-    // eoprot_config_onsay_endpoint_set(eoprot_endpoint_analogsensors, eoprot_fun_ONSAY_as);
+    // eoprot_config_onsay_endpoint_set(eoprot_endpoint_motioncontrol, eoprot_fun_ONSAY_mc);
 
     // the aim of this function is to wake up a thread which is blocked because it has sent an ask<id32>
-    // the wake up funtionality is implemented in one mode only:
-    // a. in initialisation, embObjSkin sets some values and then reads them back.
+    // the wake up funtionality is implemented in two modes, depending on the wait mechanism used:
+    // a. in initialisation, embObjMotionControl sets some values and then reads them back.
     //    the read back sends an ask<id32, signature=0xaa000000>. in such a case the board sends back
     //    a say<id32, data, signature = 0xaa000000>. thus, if the received signature is 0xaa000000, then
     //    we must unblock using feat_signal_network_reply().
+    // b. during runtime, some methods send a blocking ask<id32> without signature. It is the case of instance
+    //    of getPidRaw() which waits with a eoThreadEntry::synch() call. in such a case the board send back a
+    //    normal say<id32, data> with nos signature. in this case we unlock with wake().
+
 
     if(0xaa000000 == rd->signature)
     {   // case a:
@@ -84,36 +91,74 @@ extern void eoprot_fun_ONSAY_sk(const EOnv* nv, const eOropdescriptor_t* rd)
         {
             char str[256] = {0};
             char nvinfo[128];
-            char ipinfo[2];
             eoprot_ID2information(rd->id32, nvinfo, sizeof(nvinfo));
-            eo_common_ipv4addr_to_string(eo_nv_GetIP(nv), ipinfo, sizeof(ipinfo));
-            snprintf(str, sizeof(str), "eoprot_fun_ONSAY_sk() received an unexpected message w/ 0xaa000000 signature for IP %s and NV %s", ipinfo, nvinfo);
+            snprintf(str, sizeof(str), "eoprot_fun_ONSAY_mc() received an unexpected message w/ 0xaa000000 signature for %s", nvinfo);
             feat_PrintWarning(str);
             return;
         }
     }
-}
-
-
-
-extern void eoprot_fun_UPDT_sk_skin_status_arrayofcandata(const EOnv* nv, const eOropdescriptor_t* rd)
-{
-    EOarray* arrayof = (EOarray*)rd->data;
-    uint8_t sizeofarray = eo_array_Size(arrayof);
-    if(0 != sizeofarray)
-    {
-        feat_manage_skin_data(eo_nv_GetIP(nv), rd->id32, (void *)arrayof);
+    else
+    {   //case b:
+        wake(nv);
     }
 }
+
+
+extern void eoprot_fun_UPDT_mc_joint_status_core(const EOnv* nv, const eOropdescriptor_t* rd)
+{
+    feat_manage_motioncontrol_data(eo_nv_GetIP(nv), rd->id32, (void *)rd->data);
+}
+
+
+extern void eoprot_fun_UPDT_mc_motor_status_basic(const EOnv* nv, const eOropdescriptor_t* rd)
+{
+}
+
+
+extern void eoprot_fun_UPDT_mc_joint_status(const EOnv* nv, const eOropdescriptor_t* rd)
+{
+    feat_manage_motioncontrol_data(eo_nv_GetIP(nv), rd->id32, (void *)rd->data);
+}
+
+
+// --------------------------------------------------------------------------------------------------------------------
+// - definition of static functions
+// --------------------------------------------------------------------------------------------------------------------
+
+
+static void wake(const EOnv* nv)
+{
+    eOprotID32_t id32 = 0;
+    eOprotProgNumber_t prognum = 0 ;
+    void *mchandler = (void*) feat_MC_handler_get(eo_nv_GetIP(nv), eo_nv_GetID32(nv));
+    if(NULL == mchandler)
+    {
+        printf("eoMC class not found\n");
+        return;
+    }
+
+    id32 = eo_nv_GetID32(nv);
+    prognum = eoprot_endpoint_id2prognum(eo_nv_GetBRD(nv), id32);
+    if(eobool_false == feat_MC_mutex_post(mchandler, prognum) )
+    {
+        char nvinfo[128];
+        char ipinfo[20];
+        char str[256] = {0};
+        eoprot_ID2information(id32, nvinfo, sizeof(nvinfo));
+        eo_common_ipv4addr_to_string(eo_nv_GetIP(nv), ipinfo, sizeof(ipinfo));
+        snprintf(str, sizeof(str),"while releasing mutex for IP %s and NV %s", ipinfo, nvinfo);
+        feat_PrintWarning(str);
+    }
+
+}
+
+
+
 
 
 // --------------------------------------------------------------------------------------------------------------------
 // - end-of-file (leave a blank line after)
 // --------------------------------------------------------------------------------------------------------------------
-
-
-
-
