--- src/libraries/icubmod/embObjLib/usrcbk/EoProtocolMN_overridden_fun.h
@@ -1,72 +1,0 @@
-/*
- * Copyright (C) 2013 iCub Facility - Istituto Italiano di Tecnologia
- * Author:  Marco Accame
- * email:   marco.accame@iit.it
- * website: www.robotcub.org
- * Permission is granted to copy, distribute, and/or modify this program
- * under the terms of the GNU General Public License, version 2 or any
- * later version published by the Free Software Foundation.
- *
- * A copy of the license can be found at
- * http://www.robotcub.org/icub/license/gpl.txt
- *
- * This program is distributed in the hope that it will be useful, but
- * WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
- * Public License for more details
-*/
-
-// - include guard ----------------------------------------------------------------------------------------------------
-
-#ifndef _EOPROTOCOLMN_OVERRIDDEN_FUN_H_
-#define _EOPROTOCOLMN_OVERRIDDEN_FUN_H_
-
-
-
-
-/** @file       EoProtocolMN_overridden_fun.h
-	@brief      This header file specifies which functions are overridden in the management of management 
-	@author     marco.accame@iit.it
-	@date       06/06/2013
-**/
-
-/** @defgroup eo_asfdgr1231 Functions overridden in management endpoint
-    edfefle.
-    
-    @{		
- **/
-
-
-// this file is not needed anymore
-
-// - external dependencies --------------------------------------------------------------------------------------------
-// empty-section
-
-
-
-// - public #define  --------------------------------------------------------------------------------------------------
-
-//#define OVERRIDE_eoprot_fun_INIT_mn_comm_cmmnds_ropsigcfg
-
-
-// - declaration of public user-defined types ------------------------------------------------------------------------- 
-// empty-section
-
-    
-// - declaration of extern public variables, ... but better using use _get/_set instead -------------------------------
-// empty-section
-
-// - declaration of extern public functions ---------------------------------------------------------------------------
-// empty-section
-
-
-
-
-/** @}            
-    end of group eo_asfdgr1231  
- **/
-
-#endif  // include-guard
-
-
-// - end-of-file (leave a blank line after)----------------------------------------------------------------------------
