--- src/libraries/icubmod/embObjSkin/usrcbk/eOcfg_nvsEP_sk_overridden.h
@@ -1,36 +1,0 @@
-/*
- * Copyright (C) 2013 iCub Facility - Istituto Italiano di Tecnologia
- * Author:  Marco Accame
- * email:   marco.accame@iit.it
- * website: www.robotcub.org
- * Permission is granted to copy, distribute, and/or modify this program
- * under the terms of the GNU General Public License, version 2 or any
- * later version published by the Free Software Foundation.
- *
- * A copy of the license can be found at
- * http://www.robotcub.org/icub/license/gpl.txt
- *
- * This program is distributed in the hope that it will be useful, but
- * WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
- * Public License for more details
-*/
-
-
-// - include guard ----------------------------------------------------------------------------------------------------
-#ifndef _EOCFG_NVSEP_AS_OVERRIDDEN_H_
-#define _EOCFG_NVSEP_AS_OVERRIDDEN_H_
-
-
-// this file is to be used by the PC104, not by the EMS ...
-
-#define OVERRIDE_eo_cfg_nvsEP_sk_hid_UPDT_sstatus__arrayof10canframe
-
-
-
-#endif  // include-guard
-
-
-// eof
-
-
